+++ packages/eslint-plugin-template/src/rules/button-has-type.ts
@@ -63,7 +63,7 @@ export default createESLintRule<Options, MessageIds>({
     const parserServices = getTemplateParserServices(context);
 
     return {
-      [`Element$1[name=button]`](element: TmplAstElement) {
+      [`Element[name=button]`](element: TmplAstElement) {
         if (!isTypeAttributePresentInElement(element)) {
           if (!isIgnored(ignoreWithDirectives, element)) {
             context.report({